+++ NEWS.md
@@ -14,6 +14,7 @@
 * Add `cr_deploy_badger()` for creating build badges with Cloud Build via https://github.com/kelseyhightower/badger (#15)
 * Add `cr_deploy_run_website()` for rendering Rmd files then hosting on an nginx Cloud Run
 * The packagetools docker updates weekly `gcr.io/gcer-public/packagetools:latest` (#55)
+* Fix `cr_schedule()` crash of `overwrite=TRUE` but no existing schedule
 
 # googleCloudRunner 0.2.0